# Algorithmic improvements