--- a/lib/deePool.src.js
+++ b/lib/deePool.src.js
@@ -10,16 +10,16 @@
 })("deePool",this,function DEF(name,context){
 	"use strict";
 	
-  return { create };
+	return { create };
 
 	// ******************************
 
 	// create a new pool
 	function create(objectFactory = ()=>({})) {
 		var freePool = [];
-    var freePoolEnd = -1;
+		var freePoolEnd = -1;
 		
-    return {
+		return {
 			use,
 			recycle,
 			grow,
@@ -29,25 +29,25 @@
 		// ******************************
 
 		function use() {
-      if (freePoolEnd == -1) {
-        if (freePool.length == 0) {
-          grow(5);
-        } else {
-          grow();
-        }
-      }
-      return freePool[freePool.length - (freePoolEnd--)];
+			if (freePoolEnd == -1) {
+				if (freePool.length == 0) {
+					grow(5);
+				} else {
+					grow();
+				}
+			}
+			return freePool[freePool.length - 1 - (freePoolEnd--)];
 		}
 
 		function recycle(obj) {
-      freePool[freePool.length - (++freePoolEnd)] = obj;
+			freePool[freePool.length - (++freePoolEnd)] = obj;
 		}
 
 		function grow(count = freePool.length) {
-      for (var i = 0; i < count; ++i) {
-        freePool[freePool.length] = objectFactory();
-      }
-      freePoolEnd += count;
+			for (var i = 0; i < count; ++i) {
+				freePool[freePool.length] = objectFactory();
+			}
+			freePoolEnd += count;
 			return freePool.length;
 		}