Just some things to think about (and some poorly written example code)

The original implementation of deePool methods have an O(n) time
complexity (in particular use/recycle).
This change makes those methods O(1) and vastly simplifies the code
It makes the concession of sacrificing ordering while growing the pool
and recycling objects (this is OK as the deePool didn't specify that
there was supposed to be any ordering semantics).

Tests are still busted (due to them relying on un-specified ordering
semantics) but functionally all works.
Also forgive the tabs/spaces mismatch. 

Here's the difference in perf numbers:
Original implementation:
```
deePool Performance Benchmarks
deePool.create()... complete: 6,644,319 ops/sec (391771 samples)
use() + recycle(): small pool (10)... complete: 9,704,383 ops/sec (533343 samples)
use() + recycle(): medium pool (10,000)... complete: 14,060 ops/sec (726 samples)
use() + recycle(): large pool (1,000,000)... complete: 132 ops/sec (7 samples)
use(): growing pool... complete: 2,662,634 ops/sec (397408 samples)
grow(): small increments... complete: 4,825 ops/sec (5000 samples)
grow(): medium increments... complete: 1,395 ops/sec (150 samples)
grow(): large increments... complete: 92 ops/sec (90 samples)
grow(): exponential (doubling)... complete: 49 ops/sec (19 samples)
use() + recycle(): interleaved... complete: 1,851,973 ops/sec (203786 samples)
use() + recycle(): multiple of each... <doesn't finish in reasonable time...>
```

0(1) implementation:
```
deePool Performance Benchmarks
deePool.create()... complete: 7,929,288 ops/sec (414483 samples)
use() + recycle(): small pool... complete: 44,645,133 ops/sec (2362781 samples)
use() + recycle(): medium pool... complete: 41,821,068 ops/sec (2356688 samples)
use() + recycle(): large pool... complete: 41,899,579 ops/sec (2416356 samples)
use(): growing pool... complete: 2,555,250 ops/sec (238327 samples)
grow(): small increments... complete: 708,790 ops/sec (162082 samples)
grow(): medium increments... complete: 7,729 ops/sec (2504 samples)
grow(): large increments... complete: 367 ops/sec (90 samples)
grow(): exponential (doubling)... complete: 65 ops/sec (19 samples)
use() + recycle(): interleaved... complete: 2,192,863 ops/sec (245557 samples)
use() + recycle(): multiple of each... complete: 35,301 ops/sec (5000 samples)
```

Note the perf decrease on `use() + recycle()` as the pool gets larger in the original implementation.
The ops/sec drop is almost exactly linear with respect the increase in pool size (indicating O(n))