@joekarl 2016-12-30
Continuing conversation from twitter:
The tl;dr is this: 
- yes, `use()` is designed to be O(1) for best case, but worst case is still O(n): `use()` will attempt to traverse all elements in the pool before it can figure out that it needs to grow the pool
- `recycle()` is O(n) when `nextOpenSlot` is null (when the pool is full or when it hasn't had any objects recycled back to it or hasn't been grown with objects in use)  

This can be visualized by adding in a simple log statement to count iterations of the 'searching for free/empty slots' here https://github.com/getify/deePool/blob/master/lib/deePool.src.js#L106 and here https://github.com/getify/deePool/blob/master/lib/deePool.src.js#L57

This can also be seen in the perf numbers themselves, as the pool gets larger, the number of iterations the benchmark can run gets smaller (ie it's bound by the size of the pool).