@joekarl 2016-12-30
As a heads up, the test failures in this PR are due to tests making assumptions about ordering of recycled objects. The O(1) implementation in this PR has different ordering semantics (due to swapping pointers around the array backed pool).