@getify 2016-12-30
IIUC, this algorithm means that:

```js
var o = pool.use();
pool.recycle(o);
var p = pool.use();

o === p;   // true
```

Is that correct? This is basically a stack implementation, right?

--------

I started originally with such an algorithm, but decided not to pursue it, after working on this a bit. The reason was related to the idea that it seems like an unhealthy pool if all the churn is in one spot and other elements in the pool are potentially never being used.

By unhealthy, what I mean is that the overall performance of the app may suffer. If `use()` and `recycle()` are being interleaved (the expected normal case), `use(..)` will most of the time return an object that was just recycled, instead of (in general) returning an object that may not have ever been used yet.

Consider the cost (not at all being factored in the benchmarks) of needing to condition/reset an object. Let's say you reset every object before you reinsert. Great, you pay that cost every object, every time. Predictable. Won't differ at all between your implementation and mine.

But what if, knowing that the pool works in round-robin instead of stack, you know that for the first n times you'll pull from the pool, you're going to get an unused object each time. There, instead of reseting on recycle, you check an object after `use()` returns it, and you only reset it if it needs to be reset. With that strategy, you'll be able to avoid the cost of reset for the first n usages of the pool.

IOW, in the best case, my implementation could eek out more overall performance.

I'm not saying I don't want to go in the direction of your implementation, I'm just describing why I considered it and didn't, originally.