@getify 2016-12-30
Another thing to point out... one intention of the original algorithm was that the pool did not hold onto a reference for an object once it had been handed out... this is because, in general, it should be the freedom of the user to `use()` an object, then throw away the reference (instead of recycling it), letting it then by GC'd if they wanted. Or, instead of resetting an object, you might decide it's best to just recreate it before recycling it.

In your implementation, a reference to the original object from the pool is kept in the pool, meaning that it's possible (depending on usage pattern) that you've recreated or thrown away an object but it didn't get GC'd the way you expect.

I would prefer to have used a WeakMap, but since the performance of those is still terrible, I did the next best thing and made sure that the pool doesn't hold onto references once it hands them out.