@joekarl 2016-12-30
So I hear you on fearing the cost of resetting the objects as the travel in/out of the pool. The issue is if you're doing a round robin, assuming sufficient usage you're going to hit all of the objects in the pool anyways and end up with the same characteristic of having to reset objects all the time. And that's at the cost of having to round robin through all of your objects which means traversing the backing array (or at least part of it). 

As for reset cost, I'd bet in practice that most objects in the position of having to be pooled rarely need to be reset before used again. Take a vector or some sort of mathy object. You'll generally set all of it's member variables anyways so no reset is required. As for other objects (say game objects) these will last over the course of many frames of gameplay so the reset cost is amortized across those frames (so hardly becomes a concern).