@joekarl 2016-12-30
Ok, so the pool holding onto a reference or not is an entirely different matter, Based on the README this wasn't a feature of the lib so I considered that it was fair game.