@getify 2016-12-30
> Based on the README this wasn't a feature of the lib

Yeah, as well with the ordering, these things aren't mentioned in the README because this lib is brand new and I'm still figuring out what it should and shouldn't do. I'm describing here what original intent is, for the purposes of discussion, but I hadn't committed to any of it yet. :)