@joekarl 2016-12-30
>> Based on the README this wasn't a feature of the lib
>
>Yeah, as well with the ordering, these things aren't mentioned in the README because this lib is brand new and I'm still figuring out what it should and shouldn't do. I'm describing here what original intent is, for the purposes of discussion, but I hadn't committed to any of it yet. :)

Heh, yeah. Well sounds like you've got a plan :) Just wanted to throw out there that as currently setup this is O(n) and could easily be implemented as O(1). For a an interesting parallel, you're getting close to the idea of a [Free List](https://en.wikipedia.org/wiki/Free_list) (which I'm not sure makes tons of sense to do in JS because of the whole non-contiguous object memory thing).