@getify 2016-12-30
It's quite possible my instinct is wrong here on the idea of round-robin being better than stack. It certainly feels less "healthy" (for some definition of that word) to almost starve the pool by use/recycle staying localized to one part of the pool. But other than the potential recondition costs, I don't have any other hard analysis to back that up.

What I am pretty sure of is, the other implementations I found in other libs, they literally used the array as a stack, and did `push()` / `pop()` on it, and that kind of churn of the array is overhead I do not want the pool to pay. But a virtualized stack with a pointer is at least reasonable to consider. :)