@joekarl 2016-12-30
Yeah, so the I gather the whole point of the pool is to trade space for time (in this case memory for GC pauses). So I'm not sure you need to be concerned about having unused objects lying around. You just need enough for any point in the game (and memory is cheap for most games, even in a browser). 

And yeah, like you noted, push/pop on a real array is gonna be a bad time (well maybe... I bet the vm does some optimizations of array sizes and whatnot internally...). This is basically what one would end up doing in C or something where you're keeping track of a bunch of pointers on the heap. Is a pretty