@joekarl 2016-12-30
Anyways, just thoughts, I'll let you carry on :)