@joekarl 2016-12-30
Oh, one last thing, `recycle()` is definitely doing O(n) for normal behavior (it's really the thing contributing to the overall O(n) behavior in the perf benchmarks). So probably need to look at that at least.